Refactor: switch to md preprocessor.
Switched from a merkdown BlockProcessor to a Preprocessor.

BlockProcessor worked, but integration in an actual markdown source
revealed an incompatibility with the attr_list extension. Investigating
this in turn revealed a misunderstanding of what "block" actually means
to the markdown parser.

Switching to Preprecessing allows for a much more straightforward structure.
The incompatibility mentioned above is also fixed (added a test for it).

Went ahead and simplified a few other things like the way we hendle block
arguments.